+++ config/runtime.exs
@@ -1,0 +1,21 @@
+import Config
+
+if config_env() == :prod do
+  secret_key_base =
+    System.get_env("SECRET_KEY_BASE") ||
+      raise """
+      environment variable SECRET_KEY_BASE is missing.
+      You can generate one by calling: mix phx.gen.secret
+      """
+
+  host = System.get_env("PHX_HOST") || "localhost"
+  port = String.to_integer(System.get_env("PORT") || "8083")
+
+  config :demo_uptime, DemoUptimeWeb.Endpoint,
+    url: [host: host, port: port],
+    http: [
+      ip: {0, 0, 0, 0},
+      port: port
+    ],
+    secret_key_base: secret_key_base
+end